requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Werkzeug==2.2.2
